requirements.txt
fastapi
uvicorn
boto3
faiss-cpu
langchain